+++ noria-server/dataflow/src/ops/mod.rs
@@ -468,7 +468,7 @@ pub mod test {
 
             impl Executor for Ex {
                 fn ack(&mut self, _: SourceChannelIdentifier) {}
-                fn send_resume_at(&mut self, _: NodeIndex, _: NodeIndex, _: u32) {}
+                fn send_resume_at(&mut self, _: NodeIndex, _: NodeIndex, _: usize) {}
                 fn create_universe(&mut self, _: HashMap<String, DataType>) {}
             }
